ppt-to-md/requirements.txt
python-pptx>=1.0.0
Pillow>=10.0.0
lxml>=5.0.0
olefile>=0.47
openpyxl>=3.1.0
mammoth>=1.8.0
markdownify>=0.14.0
pymupdf>=1.24.0
pdfplumber>=0.11.0
pytesseract>=0.3.10
pytest>=8.0.0
